Added storage functions, and *not working* onPress functionality to pill button
Currently there is a bug with the styling of the button. On press, the style should toggle the background of the view containing the svg. When meds are yet to be taken, there should be no color, when they are taken, it should be green. However, currently the color stays green. However, upon inspecting the element in the app, it has the correct 'backgroundColor' style (none) applied.