added swipeable interface to month view.
There is some jitter on swiping to a new month currently. This is most likely
caused by the component having to re-render everytime the index changes. A
possible fix for this is to cache a certain number of loaded months, and also
designing a permanent loading state for the month view, so it doesn't re-render the containers, just the contained dates.